amazon: Log more helpful error when floating IP is missing
Before, if a user tried to boot a machine with an IP that wasn't
reserved, we logged the very unhelpful error:

MissingParameter: Either public IP or allocation id must be specified

This commit makes the error much clearer:

unknown floating IP 8.8.8.8. Has the IP been reserved for the region us-west-1?