ci: Increase container boot timeout
The Kibana container can take around 15 minutes to build on
DigitalOcean, so this commit increases the boot timeout for all
containers to 20 minutes.

The container takes so long to build because downloading the plugin can
take a long time (I saw 7 minutes when testing) since the plugin is
hosted in a S3 bucket in us-east, but the DigitalOcean VM is running in
SFO.

Additionally, the "optimization" step of the plugin installation takes
multiple minutes due to a Kibana implementation detail. This is true on
all of the cloud providers. See
https://discuss.elastic.co/t/x-pack-not-installed-for-kibana-6-1-0/111840
for more information.