network: Decrease chance of public connections race condition
Right now, the code that boots containers is completely disjoint from
the code that sets up inbound/outbound public network connections. The
flow of data is completely different:  the kubelet boots containers when
it reads the container from the Kubernetes-managed partition of etcd,
while the minion sets up public connections when it reads the
connections and containers from the Kelda-managed partition of etcd.
However, configuring the public connections requires the OVS ports to
have been created, so it must happen after the kubelet boots the
containers.

While this commit does not remove the possibility of this race
condition, it increases the polling interval to decrease the possible
time that the container's public connections will be unconfigured. An
upcoming change to our OVN architecture will make it possible to fully
fix this lack of synchronization between container boot and network
configuration.