cloud: Poll less aggressively once machines are booted
Before, if there were any machines running in a region, we polled it
once every five seconds. This resulted in a lot of unnecessary API
requests to the cloud provider since once the machines are booted, they
won't change unless they die.

This commit modifies the poll logic so that it polls once every 30
seconds after the machines have booted to pick up machines that have
terminated. It also tweaks the other poll intervals to take advantage of
the ability to return a specific poll interval (e.g. retrying faster
when a provider that's used by the blueprint fails to initialize).