supervisor: Run worker join regularly
Before, the supervisor join that booted the system containers on workers
only ran based on database triggers. Therefore, if booting a system container
failed (or if a system container died), the container wouldn't get
booted again until the next time the database triggered the Minion or
Etcd table.

This commit makes it so the join runs every 30 seconds, in addition to
the database triggers.