supervisor: Properly restart containers when they exit
Before, if a system container exited, the supervisor would attempt to start a
new one in its place, but the creation would fail because Docker doesn't
allow having two containers with the same name. This commit gets around
this by renaming the old container so that the new container can use the
friendly name.