cli: Make a separate base-infrastructure command
We are splitting up `kelda init` to have separate `kelda` commands for
configuring a base infrastructure and configuring cloud provider
credentials. As a first step, this commit creates a new command called
`kelda base-infrastructure` that will only create an infrastructure, and
not set up provider credentials. `kelda init` stays the same as before.