foreman: Don't update machine row when it's stopping
Before, if the foreman tried to connect to a machine while it was
stopping, it would overwrite the machine's Stopping status with a status
generated from the connection state. This was confusing because it
looked like the machine wasn't stopping anymore. Furthermore, the
information isn't useful since the machine is going to be stopped soon
anyways.